Add files via upload
The code for cleaning the data was done through Colab, and I was worried that if I did end up adding all the files to make it work in Jupiter, it would end up causing issues with my other visualizing code. And I am pretty late to seeing the requirement so I don't have the time to go over it again.